service_grpc_urgence/app/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
sqlalchemy==2.0.23
protobuf==4.25.1
